+++ java/src/main/java/com/techelevator/dao/UserDao.java
@@ -12,7 +12,7 @@ public interface UserDao {
 
     User findByUsername(String username);
 
-    int findIdByUsername(String username);
+    Long findIdByUsername(String username);
 
     boolean create(String username, String password, String role);
 }
